@@ -1,7 +1,7 @@
 [tool.black]
 line-length=88
 fast=true
-target-version=["py35", "py36", "py37", "py38"]
+target-version=["py36", "py37", "py38"]
 [tool.isort]
 multi_line_output=3
 include_trailing_comma=true
